travis上で本当にDBがリセットされているかテスト
