+++ UnitTest/BaseDbTest.cs
@@ -1,0 +1,178 @@
+﻿using Xunit;
+using UnitTest.DummyClasses;
+using System.IO;
+using System.Configuration;
+using System.Diagnostics;
+using System.Data.Common;
+using System.Text;
+
+#if ON_VISUAL_STUDIO
+using FactAttribute = Microsoft.VisualStudio.TestTools.UnitTesting.TestMethodAttribute;
+using TestClassAttribute = Microsoft.VisualStudio.TestTools.UnitTesting.TestClassAttribute;
+using ClassInitializeAttribute = Microsoft.VisualStudio.TestTools.UnitTesting.ClassInitializeAttribute;
+using ClassCleanupAttribute = Microsoft.VisualStudio.TestTools.UnitTesting.ClassCleanupAttribute;
+using TestContext = Microsoft.VisualStudio.TestTools.UnitTesting.TestContext;
+#endif
+
+using System;
+
+namespace UnitTest
+{
+  [TestClass]
+  public class BaseDbTest : BaseTest
+  {
+    public static void InitilizeClass(TestContext context)
+    {
+      ResetMySqlDatabase();
+      ResetSqlServerDatabase();
+    }
+
+    public override void FixtureSetUp()
+    {
+      BaseDbTest.InitilizeClass(null);
+    }
+
+    protected static String MySqlConnectionString
+    {
+      get { return "Server=localhost;Database=sdxtest;Uid=sdxuser;Pwd=sdx5963;"; }
+    }
+
+    protected static String SqlServerConnectionString
+    {
+      get { return "Server=.\\SQLEXPRESS;Database=sdxtest;User Id=sdxuser;Password=sdx5963;"; }
+    }
+
+    private static void ResetMySqlDatabase()
+    {
+      throw new Exception("ResetMySqlDatabase");
+      Sdx.Db.Adapter factory = new Sdx.Db.MySqlAdapter();
+
+      var masterCon = factory.CreateConnection();
+      masterCon.ConnectionString = "Server=localhost;Database=mysql;Uid=root;Pwd=";
+      using (masterCon)
+      {
+        masterCon.Open();
+
+        //drop and create db
+        try
+        {
+          var dropSql = masterCon.CreateCommand();
+          dropSql.CommandText = @"
+DROP DATABASE IF EXISTS `sdxtest` ;
+CREATE DATABASE `sdxtest` DEFAULT CHARACTER SET utf8 COLLATE utf8_general_ci;
+GRANT ALL ON `sdxtest`.* TO 'sdxuser'@'localhost' IDENTIFIED BY 'sdx5963';
+";
+          dropSql.ExecuteNonQuery();
+        }
+        catch (DbException e)
+        {
+          Console.WriteLine(e.Message);
+          //do nothing
+        }
+      }
+
+      factory.ConnectionString = BaseDbTest.MySqlConnectionString;
+      var con = factory.CreateConnection();
+      using (con)
+      {
+        con.Open();
+        BaseDbTest.ExecuteSqlFile(con, "setup.mysql.sql");
+        BaseDbTest.ExecuteSqlFile(con, "insert.sql");
+      }
+
+      Console.WriteLine("ResetMySqlDatabase");
+    }
+
+    [Conditional("ON_VISUAL_STUDIO")]
+    private static void ResetSqlServerDatabase()
+    {
+      //SdxTestデータベースをDROPします
+      Sdx.Db.Adapter factory = new Sdx.Db.SqlServerAdapter();
+      var masterCon = factory.CreateConnection();
+      String pwd = ConfigurationManager.AppSettings["SqlServerSaPwd"];
+      masterCon.ConnectionString = "Server=.\\SQLEXPRESS;Database=master;User Id=sa;Password=" + pwd;
+      using (masterCon)
+      {
+        masterCon.Open();
+
+        //drop db
+        try
+        {
+          var dropSql = masterCon.CreateCommand();
+          dropSql.CommandText = @"
+ALTER DATABASE sdxtest SET SINGLE_USER WITH ROLLBACK IMMEDIATE;
+DROP DATABASE [sdxtest]
+";
+          dropSql.ExecuteNonQuery();
+        }
+        catch (DbException e)
+        {
+          //do nothing
+        }
+
+        //drop user
+        try
+        {
+          var dropUserSql = masterCon.CreateCommand();
+          dropUserSql.CommandText = "DROP LOGIN sdxuser";
+          dropUserSql.ExecuteNonQuery();
+        }
+        catch (DbException e)
+        {
+          //do nothing
+        }
+
+        //create db
+        var createSql = masterCon.CreateCommand();
+        createSql.CommandText = "CREATE DATABASE sdxtest";
+        createSql.ExecuteNonQuery();
+
+        //create user
+        var createUserSql = masterCon.CreateCommand();
+        createUserSql.CommandText = @"
+CREATE LOGIN sdxuser WITH PASSWORD = 'sdx5963';
+ALTER AUTHORIZATION ON DATABASE::sdxtest TO sdxuser;
+";
+        createUserSql.ExecuteNonQuery();
+      }
+
+      factory.ConnectionString = BaseDbTest.SqlServerConnectionString;
+      var con = factory.CreateConnection();
+      using (con)
+      {
+        con.Open();
+        BaseDbTest.ExecuteSqlFile(con, "setup.sqlserver.sql");
+        BaseDbTest.ExecuteSqlFile(con, "insert.sql");
+      }
+
+      Console.WriteLine("ResetSqlServerDatabase");
+    }
+
+    private static void ExecuteSqlFile(DbConnection con, string dataFilePath)
+    {
+      //setup.sqlを流し込みます。
+      using (StreamReader stream = new StreamReader(dataFilePath, Encoding.GetEncoding("UTF-8")))
+      {
+        String setupSql = stream.ReadToEnd();
+        DbTransaction sqlTran = con.BeginTransaction();
+        DbCommand command = con.CreateCommand();
+        command.Transaction = sqlTran;
+
+        try
+        {
+          // Execute two separate commands.
+          command.CommandText = setupSql;
+          command.ExecuteNonQuery();
+
+          // Commit the transaction.
+          sqlTran.Commit();
+        }
+        catch (Exception ex)
+        {
+          sqlTran.Rollback();
+          throw ex;
+        }
+      }
+    }
+  }
+}
