+++ UnitTest/DbTableTest.cs
@@ -1,0 +1,71 @@
+﻿using Xunit;
+using UnitTest.DummyClasses;
+
+#if ON_VISUAL_STUDIO
+using FactAttribute = Microsoft.VisualStudio.TestTools.UnitTesting.TestMethodAttribute;
+using TestClassAttribute = Microsoft.VisualStudio.TestTools.UnitTesting.TestClassAttribute;
+using ClassInitializeAttribute = Microsoft.VisualStudio.TestTools.UnitTesting.ClassInitializeAttribute;
+using ClassCleanupAttribute = Microsoft.VisualStudio.TestTools.UnitTesting.ClassCleanupAttribute;
+using TestContext = Microsoft.VisualStudio.TestTools.UnitTesting.TestContext;
+#endif
+
+using System;
+
+namespace UnitTest
+{
+  [TestClass]
+  public class DbTableTest : BaseTest
+  {
+    [ClassInitialize]
+    public static void InitilizeClass(TestContext context)
+    {
+      Console.WriteLine("FixtureSetUp");
+      //最初のテストメソッドを実行する前に一回だけ実行したい処理はここ
+    }
+
+    [ClassCleanup]
+    public static void CleanupClass()
+    {
+      Console.WriteLine("FixtureTearDown");
+      //全てのテストメソッドが実行された後一回だけ実行する処理はここ
+    }
+
+    override protected void SetUp()
+    {
+      Console.WriteLine("SetUp");
+      //各テストメソッドの前に実行する処理はここ
+    }
+
+    override protected void TearDown()
+    {
+      Console.WriteLine("TearDown");
+      //各テストメソッドの後に実行する処理はここ
+    }
+
+    override public void FixtureSetUp()
+    {
+      TestTemplate.InitilizeClass(null);
+      //ここのクラス名は適宜書き換えてください。
+      //MSTestのFixtureSetUpがstaticじゃないとだめだったのでこのような構造になってます。
+    }
+
+    override public void FixtureTearDown()
+    {
+      TestTemplate.CleanupClass();
+      //@see FixtureSetUp
+    }
+
+
+    [Fact]
+    public void TestMethod1()
+    {
+      Console.WriteLine("TestMethod1");
+    }
+
+    [Fact]
+    public void TestMethod2()
+    {
+      Console.WriteLine("TestMethod2");
+    }
+  }
+}
